[cost_synthesis] Global state generation
In order to test instructions such as `BorrowGlobal` etc. we need to generate random valid global memory states.

This is the last diff that adds instructions to the synthesis.

Next diff(s) will refine the presentation of the instruction costs and the overall determination of the "cost" of the instruction + minor cleanup.

Note that while the `stack_generator` and `inhabitor` have similar logic in places, they can't be combined easily -- for stack generation we need to take into account such things as where the address, or bytearray is coming from, or making sure that subtraction doesn't result in overflow etc. The inhabitor doesn't need to take into account these restrictions, and simply generates an inhabitant of the given type with no constraints.