[Consensus] Pacemaker refactoring 1/n
In this diff Pacemaker stops being an EventBasedActor!

Our original motivation for adopting the EventBasedActor model for a Pacemaker was to 1) decouple "pacemaker events" generation from their processing, and 2) have a single task to modify the pacemaker state.
However, later on we introduced cases, in which event processing needed to wait for the actual processing of the events: event processing started calling async functions of a pacemaker and await for their results. As a result we lost both the advantages of the EventBasedActor model, and the code became more complicated because the Pacemaker logic was partially duplicated between the direct functions and the functionality of receiving the PacemakerEvents.
In this diff we're removing the PacemakerEvents with the corresponding channels:
* all the functions are directly invoked by EventProcessor.
* as a bonus one doesn't need to invoke a special start_event_processing_loop() function to setup the tasks. Pacemaker constructor is enough, RAII ftw.
* Pacemaker is still using an internal task responsible for processing timeouts.

The diff looks bigger than it is because multiple places depended on the EventBasedActor model. However, the only real change is in local_pacemaker.rs.
In the followup diffs I'd like to simplify the flow of events, rename some of the functions and may be remove the Pacemaker trait altogether (it's kinda silly that we have a trait with a single implementation).