[functional tests] fix resources_are_distinct_by_published_account.mvir
When this test was split into two transactions, it was changed to that the first transaction is sent by `alice` while the second one is sent by the `default` user. This is counter-intuitive and not what the test is supposed to do.

This gets it fixed so that both transactions are sent by the same account. The sequence number in the assertion needs to be bumped because we have two transactions instead of one.