[Consensus][small] SafetyRules voting does not panic
Summary:
We're planning to refactor `SafetyRules` s.t. it becomes not dependent on `BlockStore` at all,
but meanwhile we just want to not panic in case the parent is not available.
Even thought we do not fully understand the specific scenarios, in which it can happen, but it makes more sense to
return an error rather than crashing the validator.
refs #300

Test Plan: extended the unit test for consensus