[types] Add canonical serialization support for TransactionArgument
Per the title, this is a precursor to support RawTransaction.

In this diff, we reuse the TransactionArgument_ArgType, this will likely
be deleted from the code when we know longer need protobuf support for
Transactions. We will still be able to maintain consistency by matching
the behavior with new code once that happens. For now, I did not want to
litter the code base with duplicate code.

refs #454