[language] use vector in the validator set
Moving closer to making the ValidatorSet resource the source of
truth for the validators/supporting reconfiguration of consensus
by changing the validator set.

This also removes the ValidatorSet module from the genesis blob.
It should not have been there previously because it is unused.
We will restore it once it is ready to serve as the source of
truth.