[vm] fix oom discovered by fuzzer
An oom was discovered by fuzzing the interface used to deserialize a VM
Value given a VM StructDef. The oom is caused by the deserialization
logic reading a u32 from the input stream, which is intended to be the
length of a variable length vector, and attempting to pre-allocate a
vector of that length.

The fix to this oom is to:
    1) Stop attempting to pre-allocate the entire vector and instead
       allocate entries as needed.
    2) push the checks earlier which determine if this is even a valid
       `NativeStructType`.

This works in this particular case because the inputs provided by the
fuzzer are not guaranteed to be a valid (StructDef, bytes) pairs and by
the time we try to deserialize each element (of the potentially very
long vector) another deserialization error is encountered.

Closes: #1586
Approved by: dariorussi