[language] Refactor lexer to add a lookahead API
There are some cases where the parser really needs to look ahead at
the next token before deciding how to parse the current token. It was
hard to support that with the original lexer that I hacked together
from the lalrpop output, but now that the lexer is sane, it is not
so hard. Add a new lookahead API and use it in the parser, replacing
the current workarounds.

Closes: #1688
Approved by: tnowacki