[language] add Vector.remove_unstable
Support for removing elements from vectors. This PR implements the simplest and most efficient version of remove: swap the index of the element to be removed with the back of the vector, then pop the back. This is constant in the size of the vector, but may fail to preserve ordering constraints on the vector (e.g., sorting).

Closes: #1867
Approved by: sblackshear