[language][move] Minor reformatting of parser error messages
When converting from lalrpop, I tried to minimize the difference in
the parser error messages, but now that we have switched to the new
parser, I would like to make a few changes:

- Remove extra blank space after "Expected:" and before the list
of expected tokens.

- Report UnrecognizedToken errors as "Unexpected token". These tokens
are recognized perfectly well -- these problems occur when the tokens
are not expected in a particular context. The message should reflect
that.

- Use single quotes for tokens. The old parser used no quotes at all
for UnrecognizedToken errors, and it used double quotes for the list
of expected tokens. Other parts of the move-lang compiler use single
quotes, and we should be consistent.

Note that I am still hoping to quickly make more substantive improvements
in error messages, so these quick cleanups are just the first step.

Closes: #1975
Approved by: tnowacki