Supporting old on arbitrary spec expressions, and allowing implies.
This moves the `Old` variant from StorageLocation to SpecExpr in AST and parser and allows it to be used more generally (specifcally, `old(global_exists...)`).

It also adds the `==>` operator to spec expressions. This is implemented as syntactic sugar which can be isolated in the parser.

The libra_account specifications have been updated to use this new functionality, and verify more.

Closes: #2242
Approved by: sblackshear