[language/test-generation] use the Libra VM's config loading code
Previously, the test generation tool would have its own gas schedule loading
code, and was poking around in the internals of the Libra VM
(`BlockDataCache` etc) to make it happen. This was getting in the way of
Move/Libra VM separation -- we're going to move all the gas schedule loading
code into the Libra VM.

Fix this by exposing some Libra VM internals through a separate API. The
`internals` name hints at the fact that this isn't the main flow.

Closes: #2668
Approved by: runtian-zhou