[move-prover] Refactoring invariant handling, and enabling pack/unpack only model.
This PR adds a new feature allowing global spec vars to be updated just via pack/unpack invariants and without the update invariant. If a struct spec has no update invariants, then we are "synthesizing" one from the pack/unpacks s.t. if mutation starts, we treat this as unpack, and if mutation ends as pack.

The PR also cleans up invariant handling in bytecode_translator.rs which had become quite unwiedly:

- New data structure `BytecodeContext` which contains the various data which is tracked across bytecode translation of a function. More complete documentation of what is used for what.
- Factoring out some functions to reduce extensive code size of single functions (they are still too large)
- Fixed a bug in integration of lifetime analysis and simplified the handling of return. Branches and Return are special because we need to handle dead references _before_ the instruction itself, whereas for other instructions they are handled _after_. E.g. `S; goto L; <<process dead refs at S` vs `S; <<process dead refs at S>>; goto L` (later the correct one).

I have a bit more trust now that we are doing things right but it's still a complex piece  of code which we need to further improve.

Closes: #2943
Approved by: davidldill