[safety-rules] Add checks on timeout
Only sign the timeout if it is greater than or equal to the
last_voted_round and ahead of the preferred_round. We may end up signing
timeouts for rounds without first signing votes if we have received QCs
but not proposals. Always map the last_voted_round to the last signed
timeout to prevent equivocation. We can sign the last_voted_round
timeout multiple times by requiring that the underlying signing scheme
provides deterministic signatures.

Closes: #3399
Approved by: zekun000