[move-prover] Produce error if state dependent expressions are used in data invariants.
Data invariants cannot depend on global state (sender(), exists<T>, global<T>, and spec vars). This is both for methodological reasons, to prevent inconsistencies in specs, as well as for technical reasons, to let the is_wellformed type check for structures be a pure Boogie function.

In this PR we reject data invariants which depend on state during type checking. In order to do so, we need to determine whether a specification function is pure or not. Also, in the boogie translator, for specification functions which are pure, we need to not pass $m and $txn arguments. so we can continue to use spec functions in data invariants.

Closes #3511.

Closes: #3547
Approved by: DavidLDill