[CI] stop e2e test prebuild
smoketest builds the libra-node inside libra-swarm. This breaks the
original assumption that CI can prebuild the test binary and run the
tests in parallel. It is possible to prebuild the libra-node binary,
but it requires saving and restoring the build artifacts in target/
between jobs, which takes longer than building from scratch.

The container instance size was reduced recently to cut costs, but it
contributed to the e2e test flakiness. The lower amount of memory led
to occasional linker OOM which caused panic in libra-swarm.

In summary, the e2e test parallelism is reduced from 4 to 2, and the
container instance size is bumped from xlarge to 2xlarge.

Closes: #4008
Approved by: bmwill