[config] optionally retrieve peer id from storage
Look at all the places that directly access a network peer id :O
This PR moves this into secure storage so that we have one canonical
place for all tools in deployment. This will make it much easier for
deployments to access this id ranging from key-manager, management, and
the node itself so that we do not need to introduce too many dynamics
into the various config generations -- nor add more complexity to admin
overhead.

Because secure storage already depends on config, it is the natural
place to place the identity reader. I think there's a good question
about how we clean up the config structure -- but I can imagine we have
a written config that doesn't match our deployment config and we have a
tool that converts between the two... but that's future work :D.

Closes: #4181
Approved by: zekun000