[move-prover] Fixing a bug in dependency analysis.
The move-prover tries to calculate the exact set of dependencies by working with regular expressions on the source. This is better than the compiler which just puts everything in the dependency directory into the dep list, but may result in false positives via references to modules in comments. This PR filters the over-approximated deps removing any files which are sources, preventing compile errors.

Closes: #4407
Approved by: emmazzz