[libra-channel] fix race causing channel close to deadlock
We've been experiencing intermittent flaky tests in network related to
simultaneous dial tests. In the simultaneous dial logic, we sometimes
signal an existing `Peer` instance that it should close by dropping its
`libra_channel::Sender<ProtocolId, NetworkRequest>` in the `active_peers`
map. After _much_ investigation, it turns out that dropping the peer
`Sender` handle would sometimes fail to wake up the `Receiver`, who
was responsible for tearing down the `Peer` instance upon observing the
channel shutdown. As a result, the tests would sometimes deadlock,
leading to the observed flakiness in our test suite.

The fundamental issue in `libra-channel` was using the shared state
`Arc`'s `strong_count` as a proxy for the number of live `Sender`s and
the resulting race condition between dropping a `Sender` and waking the
`Receiver`.

```rust
// 1. queue is empty, one Sender, one Receiver
// 2. Sender drop
// 3. Sender acquires lock, wakes Receiver, releases lock.
// 4. Sender gets preempted _before finishing drop_ (so strong_count is
      not updated yet)
// 5. Receiver wakes up, observes no items in the queue, observes
      strong_count > 1.
// 6. Receiver registers its waker, and goes back to sleep.
// 7. Sender finally gets run and finishes dropping.
// 8. Deadlock! All Sender's are dropped and Receiver will never receive
      an external stimulus to wake again.

impl Drop for Sender {
    fn drop(&mut self) {
        // 2.
        let mut shared_state = self.shared_state.lock().unwrap();
        if let Some(waker) = shared_state.waker.take() {
            waker.wake();
        }
        // 3. compiler: shared_state unlock
        // 4. preempted
        // compiler: Arc::drop
        // 7.
    }
}

// ..

impl Stream for Receiver {
    fn poll_next(/* .. */) -> Poll<Option<Self::Item>> {
        // 5.
        let mut shared_state = self.shared_state.lock().unwrap();
        if /* popped item from queue */ {
            // ..
        } else if Arc::strong_count(&self.shared_state) == 1 {
            shared_state.stream_terminated = true;
            Poll::Ready(None)
        } else {
            // 6.
            shared_state.waker = Some(cx.waker().clone());
            Poll::Pending
        }
    }
}
```

To fix this issue, we add an explicit `num_senders` field to keep track
of the number of active `Sender` handles. When `num_senders` reaches 0
during `Sender::drop` we can safely wake the `Receiver`.

Note: unfortunately, using `num_senders` means we must acquire the lock
to increment `num_senders` on `Sender::clone`, which is intended to be a
very cheap operation... If this appears to be a bottleneck, we can try
optimizing it (ideally with the help of a concurrency model checker like
[tokio-rs/loom](https://github.com/tokio-rs/loom)).

Closes: #4479
Approved by: ankushagarwal