[consensus] add option in config to stop making progress
Add an option for consensus to only sync to last known committed state
and broadcast its own state but not vote on anything.

This helps sync every nodes to the same version under extreme situation
which needs operators to manually apply a transaction to fix the ledger
state.

Closes: #5338
Approved by: dmitri-perelman